Adicionando confirmação quando for exluir uma instituição

+++ src/pages/dashboard/InstituicaoEnsino/table/columns.tsx
@@ -13,6 +13,17 @@ import { ColumnDef } from "@tanstack/react-table";
 import { ArrowUpDown, MoreHorizontal } from "lucide-react";
 import { Link } from "react-router-dom";
 import { toast } from "sonner";
+import {
+  AlertDialog,
+  AlertDialogAction,
+  AlertDialogCancel,
+  AlertDialogContent,
+  AlertDialogDescription,
+  AlertDialogFooter,
+  AlertDialogHeader,
+  AlertDialogTitle,
+  AlertDialogTrigger,
+} from "@/components/ui/alert-dialog";
 
 export type InstituicaoEnsinoProps = {
   id: number;
@@ -90,15 +101,42 @@ export const columns: ColumnDef<InstituicaoEnsinoProps>[] = [
               <DropdownMenuItem>📝 Editar</DropdownMenuItem>
             </Link>
             <DropdownMenuSeparator />
-            <DropdownMenuItem
-              onClick={async () => {
-                toast("Instituição de Ensino exluido com Sucesso. ✅");
-                meta?.removeRow(dataRow.key);
-                await api.delete(`/instituicaoEnsino/${dataRow.id}`);
-              }}
-            >
-              🗑️ delete
-            </DropdownMenuItem>
+
+            <AlertDialog>
+              <AlertDialogTrigger asChild>
+                <Button
+                  variant="ghost"
+                  className="w-full justify-start font-normal p-2"
+                >
+                  🗑️ delete
+                </Button>
+              </AlertDialogTrigger>
+              <AlertDialogContent className="w-[90%]">
+                <AlertDialogHeader>
+                  <AlertDialogTitle>
+                    {`Deseja mesmo excluir a Instituição ${dataRow.nomeInstituicao}?`}
+                  </AlertDialogTitle>
+                  <AlertDialogDescription>
+                    Uma vez cancelada, não será possosível reverter esa ação
+                  </AlertDialogDescription>
+                </AlertDialogHeader>
+                <AlertDialogFooter>
+                  <AlertDialogCancel className="w-full">
+                    Cancel
+                  </AlertDialogCancel>
+                  <AlertDialogAction
+                    className="w-full bg-destructive text-destructive-foreground hover:bg-destructive/90"
+                    onClick={async () => {
+                      toast("Instituição de Ensino exluido com Sucesso. ✅");
+                      meta?.removeRow(dataRow.key);
+                      await api.delete(`/instituicaoEnsino/${dataRow.id}`);
+                    }}
+                  >
+                    Confirmar
+                  </AlertDialogAction>
+                </AlertDialogFooter>
+              </AlertDialogContent>
+            </AlertDialog>
           </DropdownMenuContent>
         </DropdownMenu>
       );
